pkg: move NetworkPolicy transformation code from pkg/envoy to pkg/policy/distiller
This is in preparation for a future commit which will translate rules in the
Cilium policy repository into one cilium.NetworkPolicy object per security
identity.

Signed-off by: Ian Vernon <ian@cilium.io>
