+++ pkg/policy/distiller/http.go
@@ -1,0 +1,85 @@
+// Copyright 2018 Authors of Cilium
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package distiller
+
+import (
+	"strings"
+
+	envoy_api_v2_route "github.com/cilium/cilium/pkg/envoy/envoy/api/v2/route"
+	"github.com/cilium/cilium/pkg/policy/api"
+	"github.com/cilium/cilium/pkg/policy/distiller/sort"
+	"github.com/golang/protobuf/ptypes/wrappers"
+)
+
+func GetHTTPRule(h *api.PortRuleHTTP) (headers []*envoy_api_v2_route.HeaderMatcher, ruleRef string) {
+	// Count the number of header matches we need
+	cnt := len(h.Headers)
+	if h.Path != "" {
+		cnt++
+	}
+	if h.Method != "" {
+		cnt++
+	}
+	if h.Host != "" {
+		cnt++
+	}
+
+	isRegex := wrappers.BoolValue{Value: true}
+	headers = make([]*envoy_api_v2_route.HeaderMatcher, 0, cnt)
+	if h.Path != "" {
+		headers = append(headers, &envoy_api_v2_route.HeaderMatcher{Name: ":path", Value: h.Path, Regex: &isRegex})
+		ruleRef = `PathRegexp("` + h.Path + `")`
+	}
+	if h.Method != "" {
+		headers = append(headers, &envoy_api_v2_route.HeaderMatcher{Name: ":method", Value: h.Method, Regex: &isRegex})
+		if ruleRef != "" {
+			ruleRef += " && "
+		}
+		ruleRef += `MethodRegexp("` + h.Method + `")`
+	}
+
+	if h.Host != "" {
+		headers = append(headers, &envoy_api_v2_route.HeaderMatcher{Name: ":authority", Value: h.Host, Regex: &isRegex})
+		if ruleRef != "" {
+			ruleRef += " && "
+		}
+		ruleRef += `HostRegexp("` + h.Host + `")`
+	}
+	for _, hdr := range h.Headers {
+		strs := strings.SplitN(hdr, " ", 2)
+		if ruleRef != "" {
+			ruleRef += " && "
+		}
+		ruleRef += `Header("`
+		if len(strs) == 2 {
+			// Remove ':' in "X-Key: true"
+			key := strings.TrimRight(strs[0], ":")
+			// Header presence and matching (literal) value needed.
+			headers = append(headers, &envoy_api_v2_route.HeaderMatcher{Name: key, Value: strs[1]})
+			ruleRef += key + `","` + strs[1]
+		} else {
+			// Only header presence needed
+			headers = append(headers, &envoy_api_v2_route.HeaderMatcher{Name: strs[0]})
+			ruleRef += strs[0]
+		}
+		ruleRef += `")`
+	}
+	if len(headers) == 0 {
+		headers = nil
+	} else {
+		sort.SortHeaderMatchers(headers)
+	}
+	return
+}
